障害物の実装

+++ Avoid/Avoid/Easing/easing.hpp
@@ -107,16 +107,16 @@ public:
 	public:
 		float In(float t, float b, float c, float d)	//ƒoƒO‚ ‚è
 		{
-			return -c * (std::sqrt(1 - (t /= d)*t) - 1) + b;
+			return static_cast<float>(-c * (sqrt(1 - (t /= d)*t) - 1) + b);
 		}
 		float Out(float t, float b, float c, float d)
 		{
-			return c * std::sqrt(1 - (t = t / d - 1)*t) + b;
+			return static_cast<float>(c * sqrt(1 - (t = t / d - 1)*t) + b);
 		}
 		float InOut(float t, float b, float c, float d)
 		{
-			if ((t /= d / 2) < 1) return -c / 2 * (std::sqrt(1 - t*t) - 1) + b;
-			return c / 2 * (std::sqrt(1 - t*(t -= 2)) + 1) + b;
+			if ((t /= d / 2) < 1) return static_cast<float>(-c / 2 * (sqrt(1 - t*t) - 1) + b);
+			return static_cast<float>(c / 2 * (sqrt(1 - t*(t -= 2)) + 1) + b);
 		}
 	}circ;
 	
@@ -148,7 +148,7 @@ public:
 			float a = c;
 			float s = p / 4;
 			float postFix = static_cast<float>(a*pow(2, 10 * (t -= 1)));
-			return -(postFix * sin((t*d - s)*(2 * static_cast<float>(M_PI)) / p)) + b;
+			return static_cast<float>(-(postFix * sin((t*d - s)*(2 * M_PI)) / p) + b);
 		}
 
 		float Out(float t, float b, float c, float d)
@@ -169,10 +169,10 @@ public:
 
 			if (t < 1) {
 				float postFix = static_cast<float>(a*pow(2, 10 * (t -= 1)));
-				return -.5f*(postFix* sin((t*d - s)*(2 * static_cast<float>(M_PI)) / p)) + b;
+				return static_cast<float>(-.5f*(postFix* sin((t*d - s)*(2 * M_PI)) / p) + b);
 			}
 			float postFix = static_cast<float>(a*pow(2, -10 * (t -= 1)));
-			return static_cast<float>(postFix * sin((t*d - s)*(2 * static_cast<float>(M_PI)) / p)*.5f + c + b);
+			return static_cast<float>(postFix * sin((t*d - s)*(2 * M_PI)) / p*.5f + c + b);
 		}
 	}elastic;
